ds-lfb-project-master/sql-consumer/requirements.txt
pgsql==1.2
websocket-client==1.2.1
